fix a couple more edge cases in rustdoc

+++ overlay/utils.sh
@@ -40,7 +40,6 @@ makeExternCrateFlags() {
 loadExternCrateLinkFlags() {
   local i=
   for (( i=1; i<$#; i+=2 )); do
-    local extern_name="${@:$i:1}"
     local crate="${@:((i+1)):1}"
     [ -f "$crate/.cargo-info" ] || continue
     local crate_name=`jq -r '.name' $crate/.cargo-info`
@@ -55,9 +54,10 @@ linkRustdocs() {
   local i=
   touch search-index-tmp.js
   for (( i=1; i<$#; i+=2 )); do
-    local extern_name="${@:$i:1}"
     local crate="${@:((i+1)):1}"
     [ -f "$crate/.cargo-info" ] || continue
+    # Note: The extern_name here may not match the actual package name, which is why we grab it out of the .cargo-info
+    local extern_name="$(jq -r .name $crate/.cargo-info)"
     # Dependency xyz matches self, meaning it is an older or newer version of this crate.
     # We have no way to handle this case (cargo doesn't either).
     # See https://github.com/rust-lang/cargo/issues/6313
@@ -76,6 +76,10 @@ linkRustdocs() {
     # For all dependencies we have that might appear in the generated documentation, the directory
     # containing *those* docs needs to already exist in the directory where the new docs are placed,
     # otherwise rustdoc will not generate links to the appropriate items.
+    if [ ! -e $crate/share/doc/$extern_name ]; then
+      echo >&1 "extern $crate/share/doc/$extern_name not found"
+      exit 1
+    fi
     ln -sf $crate/share/doc/$extern_name $target_dir/doc
 
     # also link up the highlighted source code. at some point, we should replace this manual stuff
@@ -85,7 +89,7 @@ linkRustdocs() {
 
     # link dependencies-of-dependencies. (anything in the share/doc directory that is a symlink
     # is a symlink to another crate's documentation)
-    find $crate/share/doc -maxdepth 1 -type l -print0 | \
+    find $crate/share/doc -maxdepth 1 -type l -not -name "$crateName" -print0 | \
       xargs -0 -i ln -Lsf {} "$target_dir/doc"
   done
   # Sort the concatenated index by crate name. GNU sort has some weirdly elaborate options, see
